core/local: Extract ./chokidar_steps/initial_diff
Including #detectOfflineUnlinkEvents() and part of #onFlush()
New .step() function may return null to keep early return in #onFlush().
